separate out get_grpc_stub
